+++ src/main/kotlin/shape/komputation/cpu/functions/Masking.kt
@@ -2,7 +2,7 @@ package shape.komputation.cpu.functions
 
 import java.util.*
 
-fun generateMask(dimension : Int, random : Random, keepProbability: Double) =
+fun generateMask(dimension : Int, random : Random, keepProbability: Float) =
 
     BooleanArray(dimension) {
