+++ src/main/kotlin/shape/komputation/cpu/layers/CpuEntryPoint.kt
@@ -1,12 +1,12 @@
 package shape.komputation.cpu.layers
 
-import shape.komputation.matrix.DoubleMatrix
+import shape.komputation.matrix.FloatMatrix
 import shape.komputation.matrix.Matrix
 
 interface CpuEntryPoint {
 
-    fun forward(input: Matrix) : DoubleMatrix
+    fun forward(input: Matrix) : FloatMatrix
 
-    fun backward(chain : DoubleMatrix) : DoubleMatrix
+    fun backward(chain : FloatMatrix) : FloatMatrix
 
 }
