+++ src/main/kotlin/shape/komputation/cpu/layers/combination/SubtractionCombination.kt
@@ -3,7 +3,7 @@ package shape.komputation.cpu.layers.combination
 import shape.komputation.cpu.functions.negate
 import shape.komputation.cpu.functions.subtract
 import shape.komputation.cpu.layers.CombinationLayer
-import shape.komputation.matrix.DoubleMatrix
+import shape.komputation.matrix.FloatMatrix
 
 /*
    Ex:
@@ -18,19 +18,19 @@ import shape.komputation.matrix.DoubleMatrix
 */
 class SubtractionCombination internal constructor(val name : String?) : CombinationLayer(name) {
 
-    override fun forward(first: DoubleMatrix, second: DoubleMatrix) =
+    override fun forward(first: FloatMatrix, second: FloatMatrix) =
 
-        DoubleMatrix(first.numberRows, first.numberColumns, subtract(first.entries, second.entries))
+        FloatMatrix(first.numberRows, first.numberColumns, subtract(first.entries, second.entries))
 
     // d (x - y) / d x = 1
-    override fun backwardFirst(chain: DoubleMatrix) =
+    override fun backwardFirst(chain: FloatMatrix) =
 
         chain
 
     // d (x - y) / d y = -1
-    override fun backwardSecond(chain: DoubleMatrix) =
+    override fun backwardSecond(chain: FloatMatrix) =
 
-        DoubleMatrix(chain.numberRows, chain.numberColumns, negate(chain.entries))
+        FloatMatrix(chain.numberRows, chain.numberColumns, negate(chain.entries))
 
 }
