+++ src/main/kotlin/shape/komputation/cpu/layers/forward/CpuColumnRepetitionLayer.kt
@@ -3,20 +3,20 @@ package shape.komputation.cpu.layers.forward
 import shape.komputation.cpu.functions.repeatColumn
 import shape.komputation.cpu.functions.sumRows
 import shape.komputation.cpu.layers.BaseCpuForwardLayer
-import shape.komputation.matrix.DoubleMatrix
+import shape.komputation.matrix.FloatMatrix
 
 class CpuColumnRepetitionLayer internal constructor(name : String? = null, private val n : Int) : BaseCpuForwardLayer(name) {
 
-    override fun forward(input : DoubleMatrix, isTraining : Boolean) : DoubleMatrix {
+    override fun forward(input : FloatMatrix, isTraining : Boolean) : FloatMatrix {
 
         val inputEntries = input.entries
         val inputSize = inputEntries.size
 
-        return DoubleMatrix(inputSize, n, repeatColumn(inputEntries, n))
+        return FloatMatrix(inputSize, n, repeatColumn(inputEntries, n))
 
     }
 
-    override fun backward(chain : DoubleMatrix): DoubleMatrix {
+    override fun backward(chain : FloatMatrix): FloatMatrix {
 
         val chainEntries = chain.entries
         val numberChainRows = chain.numberRows
@@ -24,7 +24,7 @@ class CpuColumnRepetitionLayer internal constructor(name : String? = null, priva
 
         val result = sumRows(chainEntries, numberChainRows, numberChainColumns)
 
-        return DoubleMatrix(numberChainRows, 1, result)
+        return FloatMatrix(numberChainRows, 1, result)
 
     }
