+++ src/main/kotlin/shape/komputation/cpu/layers/forward/CpuCounterProbabilityLayer.kt
@@ -3,20 +3,20 @@ package shape.komputation.cpu.layers.forward
 import shape.komputation.cpu.functions.negate
 import shape.komputation.cpu.functions.subtract
 import shape.komputation.cpu.layers.BaseCpuForwardLayer
-import shape.komputation.matrix.DoubleMatrix
+import shape.komputation.matrix.FloatMatrix
 
 class CpuCounterProbabilityLayer internal constructor(
     name : String?,
     dimension : Int) : BaseCpuForwardLayer(name) {
 
-    private val one = DoubleArray(dimension) { 1.0 }
+    private val one = FloatArray(dimension) { 1.0f }
 
-    override fun forward(input: DoubleMatrix, isTraining : Boolean) =
+    override fun forward(input: FloatMatrix, isTraining : Boolean) =
 
-        DoubleMatrix(input.numberRows, input.numberColumns, subtract(one, input.entries))
+        FloatMatrix(input.numberRows, input.numberColumns, subtract(one, input.entries))
 
-    override fun backward(chain: DoubleMatrix) =
+    override fun backward(chain: FloatMatrix) =
 
-        DoubleMatrix(chain.numberRows, chain.numberColumns, negate(chain.entries))
+        FloatMatrix(chain.numberRows, chain.numberColumns, negate(chain.entries))
 
 }
