+++ src/main/kotlin/shape/komputation/cpu/layers/forward/CpuDenseLayer.kt
@@ -3,7 +3,7 @@ package shape.komputation.cpu.layers.forward
 import shape.komputation.cpu.layers.BaseCpuForwardLayer
 import shape.komputation.cpu.layers.forward.activation.CpuActivationLayer
 import shape.komputation.cpu.layers.forward.projection.CpuProjectionLayer
-import shape.komputation.matrix.DoubleMatrix
+import shape.komputation.matrix.FloatMatrix
 import shape.komputation.optimization.Optimizable
 
 class CpuDenseLayer internal constructor(
@@ -11,7 +11,7 @@ class CpuDenseLayer internal constructor(
     private val projection : CpuProjectionLayer,
     private val activation: CpuActivationLayer) : BaseCpuForwardLayer(name), Optimizable {
 
-    override fun forward(input: DoubleMatrix, isTraining : Boolean): DoubleMatrix {
+    override fun forward(input: FloatMatrix, isTraining : Boolean): FloatMatrix {
 
         val projected = this.projection.forward(input, isTraining)
 
@@ -20,7 +20,7 @@ class CpuDenseLayer internal constructor(
         return activated
     }
 
-    override fun backward(chain: DoubleMatrix): DoubleMatrix {
+    override fun backward(chain: FloatMatrix): FloatMatrix {
 
         val diffChainWrtActivation = this.activation.backward(chain)
 
@@ -30,7 +30,7 @@ class CpuDenseLayer internal constructor(
 
     }
 
-    override fun optimize(scalingFactor : Double) {
+    override fun optimize(scalingFactor : Float) {
 
         this.projection.optimize(scalingFactor)
