+++ src/main/kotlin/shape/komputation/cpu/layers/forward/CpuTranspositionLayer.kt
@@ -2,16 +2,16 @@ package shape.komputation.cpu.layers.forward
 
 import shape.komputation.cpu.functions.transpose
 import shape.komputation.cpu.layers.BaseCpuForwardLayer
-import shape.komputation.matrix.DoubleMatrix
+import shape.komputation.matrix.FloatMatrix
 
 class CpuTranspositionLayer internal constructor(name : String? = null) : BaseCpuForwardLayer(name)  {
 
-    override fun forward(input: DoubleMatrix, isTraining : Boolean) =
+    override fun forward(input: FloatMatrix, isTraining : Boolean) =
 
-        DoubleMatrix(input.numberColumns, input.numberRows, transpose(input.numberRows, input.numberColumns, input.entries))
+        FloatMatrix(input.numberColumns, input.numberRows, transpose(input.numberRows, input.numberColumns, input.entries))
 
-    override fun backward(chain: DoubleMatrix) =
+    override fun backward(chain: FloatMatrix) =
 
-        DoubleMatrix(chain.numberColumns, chain.numberRows, transpose(chain.numberRows, chain.numberColumns, chain.entries))
+        FloatMatrix(chain.numberColumns, chain.numberRows, transpose(chain.numberRows, chain.numberColumns, chain.entries))
 
 }
