+++ src/main/kotlin/shape/komputation/cpu/layers/forward/activation/CpuIdentityLayer.kt
@@ -1,14 +1,14 @@
 package shape.komputation.cpu.layers.forward.activation
 
-import shape.komputation.matrix.DoubleMatrix
+import shape.komputation.matrix.FloatMatrix
 
 class CpuIdentityLayer internal constructor(name : String? = null) : BaseCpuActivationLayer(name) {
 
-    override fun forward(input : DoubleMatrix, isTraining : Boolean) =
+    override fun forward(input : FloatMatrix, isTraining : Boolean) =
 
         input
 
-    override fun backward(chain : DoubleMatrix) =
+    override fun backward(chain : FloatMatrix) =
 
         chain
